Split game_core into core.game package with shims
