+++ src/core/game/__init__.py
@@ -1,0 +1,7 @@
+"""Game logic package."""
+
+from .engine import GameCore, EnhancedGameCore, create_enhanced_game
+from .state import GameState
+
+__all__ = ["GameCore", "EnhancedGameCore", "GameState", "create_enhanced_game"]
+
