+++ src/core/game/cultivation.py
@@ -1,0 +1,6 @@
+"""Cultivation related helpers."""
+
+
+class CultivationMixin:
+    def _do_cultivate(self) -> None:  # pragma: no cover - simplified
+        pass
